Refactor catalog files and unit tests.
Closes #785
