+++ tests/test_CanFam.py
@@ -1,31 +1,18 @@
-import os
-import unittest
-from unittest import mock
-
 import stdpopsim
-import stdpopsim.cli
 from tests import test_species
 
 
-class TestSpecies(unittest.TestCase, test_species.SpeciesTestMixin):
+class TestSpecies(test_species.SpeciesTestBase):
     species = stdpopsim.get_species("CanFam")
 
     def test_basic_attributes(self):
-        self.assertEqual(self.species.population_size, 13000)
-        self.assertEqual(self.species.generation_time, 3)
+        assert self.species.population_size == 13000
+        assert self.species.generation_time == 3
 
 
-class TestGenome(unittest.TestCase, test_species.GenomeTestMixin):
+class TestGenome(test_species.GenomeTestBase):
     genome = stdpopsim.get_species("CanFam").genome
 
     def test_basic_attributes(self):
         nchrom = 40  # 38 + X + MT
-        self.assertEqual(len(self.genome.chromosomes), nchrom)
-
-
-class TestThatDogsCanBeSimulated(unittest.TestCase):
-    def test_basic_cli_usage(self):
-        cmd = "-q CanFam -c 38 -l 0.001 --seed 1234 10"
-        with mock.patch("sys.stdout", autospec=True) as stdout:
-            stdout.buffer = open(os.devnull, "wb")
-            stdpopsim.cli.stdpopsim_main(cmd.split())
+        assert len(self.genome.chromosomes) == nchrom
