+++ tests/test_DroMel.py
@@ -1,21 +1,21 @@
 """
 Tests for the drosophila_melanogaster data definitions.
 """
-import unittest
+import pytest
 
 import stdpopsim
 from tests import test_species
 
 
-class TestSpecies(unittest.TestCase, test_species.SpeciesTestMixin):
+class TestSpecies(test_species.SpeciesTestBase):
     species = stdpopsim.get_species("DroMel")
 
     def test_basic_attributes(self):
-        self.assertEqual(self.species.population_size, 1720600)
-        self.assertEqual(self.species.generation_time, 0.1)
+        self.species.population_size == 1720600
+        self.species.generation_time == 0.1
 
 
-class TestGenome(unittest.TestCase, test_species.GenomeTestMixin):
+class TestGenome(test_species.GenomeTestBase):
     """
     Tests for the drosophila_melanogaster genome.
     """
@@ -23,15 +23,14 @@ class TestGenome(unittest.TestCase, test_species.GenomeTestMixin):
     genome = stdpopsim.get_species("DroMel").genome
 
     def test_basic_attributes(self):
-        self.assertEqual(len(self.genome.chromosomes), 8)
+        assert len(self.genome.chromosomes) == 8
 
-    def test_non_recombining_chrs(self):
-        chrom = self.genome.get_chromosome("Y")
-        self.assertEqual(chrom.recombination_rate, 0)
-        chrom = self.genome.get_chromosome("mitochondrion_genome")
-        self.assertEqual(chrom.recombination_rate, 0)
+    @pytest.mark.parametrize("chr_id", ["Y", "mitochondrion_genome"])
+    def test_non_recombining_chrs(self, chr_id):
+        chrom = self.genome.get_chromosome(chr_id)
+        assert chrom.recombination_rate == 0
 
-    def test_recombining_chrs(self):
-        for name in ["2L", "2R", "3L", "3R", "4", "X"]:
-            chrom = self.genome.get_chromosome(name)
-            self.assertGreater(chrom.recombination_rate, 0)
+    @pytest.mark.parametrize("chr_id", ["2L", "2R", "3L", "3R", "4", "X"])
+    def test_recombining_chrs(self, chr_id):
+        chrom = self.genome.get_chromosome(chr_id)
+        assert chrom.recombination_rate > 0
