+++ tests/test_EscCol.py
@@ -1,25 +1,23 @@
 """
 Tests for the e. coli data definitions.
 """
-import unittest
-
 import stdpopsim
 from tests import test_species
 
 
-class TestSpecies(unittest.TestCase, test_species.SpeciesTestMixin):
+class TestSpecies(test_species.SpeciesTestBase):
     species = stdpopsim.get_species("EscCol")
 
     def test_basic_attributes(self):
         # From paper https://doi.org/10.1093/molbev/msw048
         # Ne taken from Table 2
-        self.assertEqual(self.species.population_size, 1.8e8)
+        assert self.species.population_size == 1.8e8
         # 20 minutes per generation
         generation_time = 1.0 / (525600 / 20)
-        self.assertAlmostEqual(self.species.generation_time, generation_time)
+        assert round(abs(self.species.generation_time - generation_time), 7) == 0
 
 
-class TestGenome(unittest.TestCase, test_species.GenomeTestMixin):
+class TestGenome(test_species.GenomeTestBase):
     """
     Tests for the e_coli genome.
     """
@@ -27,4 +25,4 @@ class TestGenome(unittest.TestCase, test_species.GenomeTestMixin):
     genome = stdpopsim.get_species("EscCol").genome
 
     def test_basic_attributes(self):
-        self.assertEqual(len(self.genome.chromosomes), 1)
+        assert len(self.genome.chromosomes) == 1
