+++ tests/test_HomSap.py
@@ -1,44 +1,48 @@
 """
 Tests for the human data definitions.
 """
-import unittest
-
 import stdpopsim
+
+import pytest
 
 from tests import test_species
 
 
-class TestSpecies(unittest.TestCase, test_species.SpeciesTestMixin):
+class TestSpecies(test_species.SpeciesTestBase):
     species = stdpopsim.get_species("HomSap")
 
     def test_basic_attributes(self):
-        self.assertEqual(self.species.population_size, 10 ** 4)
-        self.assertEqual(self.species.generation_time, 30)
+        assert self.species.population_size == 10 ** 4
+        assert self.species.generation_time == 30
 
 
-class TestGenome(unittest.TestCase, test_species.GenomeTestMixin):
-    """
-    Tests for the human genome.
-    """
+class TestGenome(test_species.GenomeTestBase):
 
     genome = stdpopsim.get_species("HomSap").genome
 
     def test_basic_attributes(self):
-        self.assertEqual(len(self.genome.chromosomes), 25)
+        assert len(self.genome.chromosomes) == 25
 
-    def test_recombination_rates(self):
+    @pytest.mark.parametrize("chr_id", [chrom.id for chrom in genome.chromosomes])
+    def test_recombination_rates(self, chr_id):
         # recompute recombination rates from HapMapII_GRCh37 map then
         # compare the results to the current recombination rates for each chromosome
         genetic_map = "HapMapII_GRCh37"
         species = stdpopsim.get_species("HomSap")
-        for chrom in self.genome.chromosomes:
-            if chrom.id == "Y":
-                with self.assertWarns(Warning):
-                    contig = species.get_contig(chrom.id, genetic_map=genetic_map)
-                print("HERE")
-            else:
-                contig = species.get_contig(chrom.id, genetic_map=genetic_map)
-            self.assertAlmostEqual(
-                chrom.recombination_rate,
-                contig.recombination_map.mean_rate,
-            )
+        chrom = species.genome.get_chromosome(chr_id)
+        if chr_id in ["X", "Y", "MT"]:
+            with pytest.warns(stdpopsim.NonAutosomalWarning):
+                contig = species.get_contig(chr_id, genetic_map=genetic_map)
+        elif chr_id in ["3", "5", "7", "11", "16", "17", "18", "20"]:
+            contig = species.get_contig(chr_id, genetic_map=genetic_map)
+        else:
+            # The rest of the chromosomes are currently emitting a warning about
+            # the mismatch in chromosome lengths because of the fact that we're
+            # on 37 for the map. This should be resolved when we start using the
+            # lifted over map.
+            with pytest.warns(UserWarning, match="longer than chromosome length"):
+                contig = species.get_contig(chr_id, genetic_map=genetic_map)
+        assert pytest.approx(
+            chrom.recombination_rate,
+            contig.recombination_map.mean_rate,
+        )
